chore(testing-drivers): refactor testing shared modules
