+++ packages/cubejs-testing-drivers/src/dataset.ts
@@ -1,17 +1,4 @@
-export type Cast = {
-  DATE_PREFIX: string,
-  DATE_SUFFIX: string,
-  SELECT_PREFIX: string,
-  SELECT_SUFFIX: string,
-  CREATE_TBL_PREFIX: string,
-  CREATE_TBL_SUFFIX: string,
-  CREATE_SUB_PREFIX: string,
-  CREATE_SUB_SUFFIX: string,
-  USE_SCHEMA: string,
-  GENERATE_BIG_SERIES?: string,
-  TRUE_LITERAL?: string,
-  FALSE_LITERAL?: string,
-};
+import { Cast } from './types/Cast';
 
 function create(table: string, query: string, cast: Cast, suf?: string): string {
   return `
