+++ packages/cubejs-testing-shared/src/db-container-runners/db-runner.abstract.ts
@@ -5,7 +5,8 @@ export interface ContainerVolumeDefinition {
 }
 
 export interface DBRunnerContainerOptions {
-  volumes?: ContainerVolumeDefinition[]
+  volumes?: ContainerVolumeDefinition[],
+  version?: string,
 }
 
 // @todo Finish with old one PR for it
