+++ packages/cubejs-testing-shared/src/db-container-runners/prestodb.ts
@@ -2,15 +2,11 @@ import { GenericContainer, Wait } from 'testcontainers';
 
 import { DbRunnerAbstract, DBRunnerContainerOptions } from './db-runner.abstract';
 
-type TrinoStartOptions = DBRunnerContainerOptions & {
-    version?: string,
-};
+export class PrestoDbRunner extends DbRunnerAbstract {
+  public static startContainer(options: DBRunnerContainerOptions) {
+    const version = process.env.TEST_PRESTO_VERSION || options.version || '0.281';
 
-export class TrinoDBRunner extends DbRunnerAbstract {
-  public static startContainer(options: TrinoStartOptions) {
-    const version = process.env.TEST_TRINO_VERSION || options.version || '403';
-
-    const container = new GenericContainer(`trinodb/trino:${version}`)
+    const container = new GenericContainer(`ahanaio/prestodb-sandbox:${version}`)
       .withExposedPorts(8080)
       .withWaitStrategy(Wait.forLogMessage('======== SERVER STARTED ========'))
       .withStartupTimeout(30 * 1000);
