+++ packages/cubejs-testing-shared/src/index.ts
@@ -4,6 +4,6 @@ export * from './query-test.abstract';
 export * from './DriverTests';
 export * from './utils';
 
-export * from './db';
+export * from './db-container-runners';
 
 export const smartStringTrim = dedent;
